Add option -t to test a configuration
In contrast to some other JSON parsers which might understand relaxed
JSON, the parser we use wants correct JSON format. And JSON is quiet
error prone in regard to the use of commas.

Therefor add an option -t to make it easy to test a configuration before
really using it.